Updated columns sizes

+++ config/database.sql
@@ -23,7 +23,7 @@ USE webframework;
 // Create the tables
 // ****************************************************************************/
 CREATE TABLE user_status (
-    id INT(10) UNSIGNED NOT NULL AUTO_INCREMENT,
+    id TINYINT(1) UNSIGNED NOT NULL AUTO_INCREMENT,
     
     status VARCHAR(25) NOT NULL,
     
@@ -42,7 +42,7 @@ CREATE TABLE user (
     email VARCHAR(100) NOT NULL,
     password CHAR(60) NOT NULL,
     
-    status_id INT(1) UNSIGNED NOT NULL DEFAULT 1,
+    status_id TINYINT(1) UNSIGNED NOT NULL DEFAULT 1,
     
     created_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
